requirements.txt
numpy
torch
torchvision
opencv-python
Pillow
datasets
transformers
peft
trl
pandas
tqdm
scikit-learn
